backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
Werkzeug
psycopg2-binary
google-auth
google-auth-oauthlib
